remove redistributer change
